+++ src/Stormancer.Plugins/Queries/Stormancer.Server.Plugins.Queries/IServiceSearchProvider.cs
@@ -7,6 +7,6 @@ namespace Stormancer.Server.Plugins.Queries
     public interface IServiceSearchProvider
     {
         bool Handles(string type);
-        IEnumerable<Document<JObject>> Filter(string type, JObject filter, uint size);
+        SearchResult<JObject> Filter(string type, JObject filter, uint size);
     }
 }