+++ src/Stormancer.Plugins/SocketApi/Stormancer.Server.Plugins.SocketApi/Changelog.rst
@@ -12,5 +12,6 @@ This project adheres to semantic versioning.
 Unreleased
 ----------
 Added
-*******
+*****
 - Added plugin. The SocketApi plugin contains systems enabling clients to benefit from an hybrid relay/P2P direct communication API between clients on top of scenes, that mimicks the socket "Send/receive" API model. The API is used by the Stormancer Unreal Engine plugin to power the StormancerSocketSubsystem.
+